+++ src/main/scala/ru/ifmo/onell/util/IntSet.scala
@@ -4,9 +4,9 @@ import java.util.Random
 
 trait IntSet {
   def size: Int
-  def test(element: Int): Boolean
-  def apply(index: Int): Int
+  def test(element: Long): Boolean
+  def apply(index: Int): Long
   def clear(): Unit
-  def add(element: Int): Unit
+  def add(element: Long): Unit
   def shuffleOrder(rng: Random): Unit
 }
